Move some getter functions from revision.go to cruds.go
